+++ packages/react-native/src/icons/index.ts
@@ -2,4 +2,3 @@ import * as AppIcons from "./app"
 import * as ModuleIcons from "./modules"
 export { AppIcons, ModuleIcons }
 export * from "./types"
-export { Icon } from "../components/Icon"